refactor: update cs proj resolve runner to use kd_clients
