+++ deps.ts
@@ -1,15 +1,19 @@
+import { existsSync } from "https://deno.land/std@0.207.0/fs/exists.ts";
 import {
 	IssueClient, PullRequestClient, ProjectClient,
 	OrgClient, RepoClient, TagClient, LabelClient, MilestoneClient, GitClient
 } from "https://deno.land/x/kd_clients@v1.0.0-preview.4/GitHubClients/mod.ts";
+import { XClient } from "https://deno.land/x/kd_clients@v1.0.0-preview.4/OtherClients/mod.ts";
 import { Input } from "https://deno.land/x/cliffy@v1.0.0-rc.3/prompt/input.ts";
 import chalk from "npm:chalk@5.3.0";
-import { Directory, CLI, File } from "./cicd/core/mod.ts";
+import { Directory, CLI, File, Path } from "./cicd/core/mod.ts";
 
+export { existsSync };
 export {
-	IssueClient, PullRequestClient, ProjectClient,
-	OrgClient, RepoClient, TagClient, LabelClient, MilestoneClient, GitClient
+	IssueClient, PullRequestClient, ProjectClient, MilestoneClient,
+	OrgClient, RepoClient, TagClient, LabelClient, GitClient,
+	XClient
 };
 export { Input };
 export default chalk;
-export { Directory, CLI, File };
+export { Directory, CLI, File, Path };
